+++ flamenco/settings.go
@@ -45,18 +45,118 @@ const (
 	configFilename   = "flamenco-manager.yaml"
 	defaultServerURL = "https://cloud.blender.org/"
 
-	// relative to the Flamenco Server base URL:
+	latestConfigVersion = 2
+
+	// relative to the Flamenco Server Base URL:
 	jwtPublicKeysRelativeURL = "api/flamenco/jwt/public-keys"
 )
 
 var (
-	// ErrDuplicateVariables is returned when the same name is used as regular and path-replacement variable.
-	ErrDuplicateVariables = errors.New("duplicate variables found")
+	// ErrMissingVariablePlatform is returned when a variable doesn't declare any valid platform for a certain value.
+	ErrMissingVariablePlatform = errors.New("variable's value is missing platform declaration")
+	ErrBadDirection            = errors.New("variable's direction is invalid")
 
 	// Valid values for the "mode" config variable.
 	validModes = map[string]bool{
 		"develop":    true,
 		"production": true,
+	}
+
+	// Valid values for the "audience" tag of a ConfV2 variable.
+	validAudiences = map[string]bool{
+		"":        true,
+		"all":     true,
+		"workers": true,
+		"users":   true,
+	}
+
+	// The default configuration, use DefaultConfig() to obtain a copy.
+	defaultConfig = Conf{
+		Base: Base{
+			Meta: ConfMeta{Version: latestConfigVersion},
+
+			Mode:                        "production",
+			ManagerName:                 "Flamenco Manager",
+			Listen:                      ":8083",
+			DatabasePath:                "./db",
+			TaskLogsPath:                "./task-logs",
+			DownloadTaskSleep:           10 * time.Minute,
+			DownloadTaskRecheckThrottle: 10 * time.Second,
+			TaskUpdatePushMaxInterval:   5 * time.Second,
+			TaskUpdatePushMaxCount:      3000,
+			CancelTaskFetchInterval:     10 * time.Second,
+			ActiveTaskTimeoutInterval:   10 * time.Minute,
+			ActiveWorkerTimeoutInterval: 1 * time.Minute,
+			FlamencoStr:                 defaultServerURL,
+			// Days are assumed to be 24 hours long. This is not exactly accurate, but should
+			// be accurate enough for this type of cleanup.
+			TaskCleanupMaxAge: 14 * 24 * time.Hour,
+			SSDPDiscovery:     true,
+			SSDPDeviceUUID:    "7401c189-ef69-434b-b4d8-56d00075faf5",
+
+			BlacklistThreshold:         3,
+			TaskFailAfterSoftFailCount: 3,
+
+			WorkerCleanupStatus: []string{workerStatusOffline},
+
+			TestTasks: TestTasks{
+				BlenderRender: BlenderRenderConfig{
+					JobStorage:   "{job_storage}/test-jobs",
+					RenderOutput: "{render}/test-renders",
+				},
+			},
+
+			Shaman: shamanconfig.Config{
+				FileStorePath: "../shaman-file-store",
+				CheckoutPath:  "../shaman-checkout",
+
+				GarbageCollect: shamanconfig.GarbageCollect{
+					Period:            0,
+					MaxAge:            31 * 24 * time.Hour,
+					ExtraCheckoutDirs: []string{},
+				},
+			},
+
+			JWT: jwtauth.Config{
+				DownloadKeysInterval: 1 * time.Hour,
+			},
+		},
+
+		Variables: map[string]ConfV2Variable{
+			"blender": ConfV2Variable{
+				Direction: "oneway",
+				Values: ConfV2VariableValues{
+					ConfV2VariableValue{Platform: "linux", Audience: "users", Value: "/linux/path/to/blender"},
+					ConfV2VariableValue{Platform: "linux", Audience: "workers", Value: "/farm/path/to/blender"},
+					ConfV2VariableValue{Platform: "windows", Value: "C:/windows/path/to/blender.exe"},
+					ConfV2VariableValue{Platform: "darwin", Value: "/Volumes/Applications/Blender/blender"},
+				},
+			},
+			"ffmpeg": ConfV2Variable{
+				Direction: "oneway",
+				Values: ConfV2VariableValues{
+					ConfV2VariableValue{Platform: "linux", Value: "/usr/bin/ffmpeg"},
+					ConfV2VariableValue{Platform: "windows", Value: "C:/windows/path/to/ffmpeg.exe"},
+					ConfV2VariableValue{Platform: "darwin", Value: "/Volumes/Applications/FFmpeg/ffmpeg"},
+				},
+			},
+			"render": ConfV2Variable{
+				Direction: "twoway",
+				Values: ConfV2VariableValues{
+					ConfV2VariableValue{Platform: "linux", Value: "/render"},
+					ConfV2VariableValue{Platform: "windows", Value: "R:"},
+					ConfV2VariableValue{Platform: "darwin", Value: "/Volumes/render"},
+				},
+			},
+			"job_storage": ConfV2Variable{
+				Direction: "twoway",
+				Values: ConfV2VariableValues{
+					ConfV2VariableValue{Platform: "linux", Value: "/shared/flamenco-jobs"},
+					ConfV2VariableValue{Platform: "windows", Value: "S:"},
+					ConfV2VariableValue{Platform: "darwin", Value: "/Volumes/shared/flamenco-jobs"},
+				},
+			},
+		},
 	}
 )
 
@@ -71,8 +171,16 @@ type TestTasks struct {
 	BlenderRender BlenderRenderConfig `yaml:"test_blender_render"`
 }
 
-// Conf represents the Manager's configuration file.
-type Conf struct {
+// ConfMeta contains configuration file metadata.
+type ConfMeta struct {
+	// Version of the config file structure.
+	Version int `yaml:"version"`
+}
+
+// Base contains those settings that are shared by all configuration versions.
+type Base struct {
+	Meta ConfMeta `yaml:"_meta"`
+
 	Mode          string   `yaml:"mode"` // either "develop" or "production"
 	ManagerName   string   `yaml:"manager_name"`
 	DatabaseURL   string   `yaml:"database_url"`
@@ -83,7 +191,7 @@ type Conf struct {
 	FlamencoStr   string   `yaml:"flamenco"`
 	Flamenco      *url.URL `yaml:"-"`
 	ManagerID     string   `yaml:"manager_id"`
-	ManagerSecret string   `yaml:"manager_secret"`
+	ManagerSecret string   `yaml:"manager_secret,omitempty"`
 	TLSKey        string   `yaml:"tlskey"`
 	TLSCert       string   `yaml:"tlscert"`
 
@@ -94,16 +202,6 @@ type Conf struct {
 	 * If set to -1, will never check when a worker asks for a task (so only every
 	 * download_task_sleep_seconds seconds). */
 	DownloadTaskRecheckThrottle time.Duration `yaml:"download_task_recheck_throttle"`
-
-	/* Variables, stored differently in YAML and these settings.
-	 * Variables:             variable name -> platform -> value
-	 * VariablesPerPlatform:  platform -> variable name -> value
-	 */
-	VariablesByVarname  map[string]map[string]string `yaml:"variables"`
-	VariablesByPlatform map[string]map[string]string `yaml:"-"`
-
-	PathReplacementByVarname  map[string]map[string]string `yaml:"path_replacement"`
-	PathReplacementByPlatform map[string]map[string]string `yaml:"-"`
 
 	TaskUpdatePushMaxInterval time.Duration `yaml:"task_update_push_max_interval"`
 	TaskUpdatePushMaxCount    int           `yaml:"task_update_push_max_count"`
@@ -139,109 +237,257 @@ type Conf struct {
 	WorkerRegistrationSecret string         `yaml:"worker_registration_secret"`
 }
 
+// Conf is the latest version of the configuration.
+// Currently it is version 2.
+type Conf struct {
+	Base `yaml:",inline"`
+
+	// Variable name → Variable definition
+	Variables map[string]ConfV2Variable `yaml:"variables"`
+
+	// audience + platform + variable name → variable value.
+	// Used to look up variables for a given platform and audience.
+	// The 'audience' is never "all" or ""; only concrete audiences are stored here.
+	VariablesLookup map[string]map[string]map[string]string `yaml:"-"`
+}
+
+// ConfV2Variable defines a variable (in config version 2)
+type ConfV2Variable struct {
+	// Either "oneway" or "twoway"
+	Direction string `yaml:"direction" json:"direction"`
+	// Mapping from variable value to audience/platform definition.
+	Values ConfV2VariableValues `yaml:"values" json:"values"`
+}
+
+// ConfV2VariableValues is the list of values of a variable.
+type ConfV2VariableValues []ConfV2VariableValue
+
+// ConfV2VariableValue defines which audience and platform see which value.
+type ConfV2VariableValue struct {
+	// Audience defines who will use this variable, either "all", "workers", or "users". Empty string is "all".
+	Audience string `yaml:"audience,omitempty" json:"audience,omitempty"`
+
+	// Platforms that use this value. Only one of "Platform" and "Platforms" may be set.
+	Platform  string   `yaml:"platform,omitempty" json:"platform,omitempty"`
+	Platforms []string `yaml:"platforms,omitempty,flow" json:"platforms,omitempty,flow"`
+
+	// The actual value of the variable for this audience+platform.
+	Value string `yaml:"value" json:"value"`
+}
+
 // GetConf parses flamenco-manager.yaml and returns its contents as a Conf object.
 func GetConf() (Conf, error) {
 	return LoadConf(configFilename)
 }
 
+// DefaultConfig returns a copy of the default configuration.
+func DefaultConfig() Conf {
+	c := defaultConfig
+	c.Meta.Version = latestConfigVersion
+	c.constructVariableLookupTable()
+	return c
+}
+
 // LoadConf parses the given file and returns its contents as a Conf object.
 func LoadConf(filename string) (Conf, error) {
 	yamlFile, err := ioutil.ReadFile(filename)
-
-	// Construct the struct with some more or less sensible defaults.
-	c := Conf{
-		Mode:                        "production",
-		ManagerName:                 "Flamenco Manager",
-		Listen:                      ":8083",
-		DatabasePath:                "./db",
-		TaskLogsPath:                "./task-logs",
-		DownloadTaskSleep:           10 * time.Minute,
-		DownloadTaskRecheckThrottle: 10 * time.Second,
-		TaskUpdatePushMaxInterval:   5 * time.Second,
-		TaskUpdatePushMaxCount:      3000,
-		CancelTaskFetchInterval:     10 * time.Second,
-		ActiveTaskTimeoutInterval:   10 * time.Minute,
-		ActiveWorkerTimeoutInterval: 1 * time.Minute,
-		FlamencoStr:                 defaultServerURL,
-		// Days are assumed to be 24 hours long. This is not exactly accurate, but should
-		// be accurate enough for this type of cleanup.
-		TaskCleanupMaxAge: 14 * 24 * time.Hour,
-		SSDPDiscovery:     true,
-		SSDPDeviceUUID:    "7401c189-ef69-434b-b4d8-56d00075faf5",
-
-		BlacklistThreshold:         3,
-		TaskFailAfterSoftFailCount: 3,
-
-		WorkerCleanupStatus: []string{workerStatusOffline},
-
-		VariablesByVarname: map[string]map[string]string{
-			"blender": map[string]string{
-				"linux":   "/linux/path/to/blender",
-				"windows": "C:/windows/path/to/blender.exe",
-				"darwin":  "/Volume/Applications/Blender/blender",
-			},
-			"ffmpeg": map[string]string{
-				"linux":   "/usr/bin/ffmpeg",
-				"windows": "C:/windows/path/to/ffmpeg.exe",
-				"darwin":  "/Volume/Applications/FFmpeg/ffmpeg",
-			},
-		},
-
-		PathReplacementByVarname: map[string]map[string]string{
-			"render": map[string]string{
-				"linux":   "/render",
-				"windows": "R:/",
-				"darwin":  "/Volume/render",
-			},
-			"job_storage": map[string]string{
-				"linux":   "/shared/flamenco-jobs",
-				"windows": "S:/",
-				"darwin":  "/Volume/shared/flamenco-jobs",
-			},
-		},
-
-		TestTasks: TestTasks{
-			BlenderRender: BlenderRenderConfig{
-				JobStorage:   "{job_storage}/test-jobs",
-				RenderOutput: "{render}/test-renders",
-			},
-		},
-
-		Shaman: shamanconfig.Config{
-			FileStorePath: "../shaman-file-store",
-			CheckoutPath:  "../shaman-checkout",
-
-			GarbageCollect: shamanconfig.GarbageCollect{
-				Period:            0,
-				MaxAge:            31 * 24 * time.Hour,
-				ExtraCheckoutDirs: []string{},
-			},
-		},
-
-		JWT: jwtauth.Config{
-			DownloadKeysInterval: 1 * time.Hour,
-		},
-	}
 	if err != nil {
-		c.processVariables()
-		return c, err
+		return DefaultConfig(), err
 	}
 
-	err = yaml.Unmarshal(yamlFile, &c)
-	if err != nil {
-		c.processVariables()
-		return c, fmt.Errorf("unmarshal: %v", err)
+	// First parse attempt, find the version.
+	baseConf := Base{}
+	if err := yaml.Unmarshal(yamlFile, &baseConf); err != nil {
+		log.WithError(err).Fatal("unable to parse YAML as Base config")
+	}
+	logger := log.WithField("version", baseConf.Meta.Version)
+	if baseConf.Meta.Version < latestConfigVersion {
+		logger = logger.WithField("latestVersion", latestConfigVersion)
 	}
 
+	// Second parse attempt, based on the version found.
+	c := DefaultConfig()
+	switch baseConf.Meta.Version {
+	case 0, 1:
+		logger.Info("upgrading config to version 2")
+		confv1 := defaultConfigV1()
+		if err := yaml.Unmarshal(yamlFile, &confv1); err != nil {
+			logger.WithError(err).Fatal("unable to parse YAML")
+		}
+		confv1.upgradeToV2(&c)
+	case 2:
+		log.Info("interpreting as version 2")
+		if err := yaml.Unmarshal(yamlFile, &c); err != nil {
+			logger.WithError(err).Fatal("unable to parse YAML")
+		}
+	}
+
+	c.constructVariableLookupTable()
 	c.parseURLs()
 	c.checkMode(c.Mode)
 	c.checkDatabase()
+	c.checkVariables()
 
-	err = c.processVariables()
-	if err != nil {
-		return c, err
-	}
 	return c, nil
+}
+
+func (c *Conf) constructVariableLookupTable() {
+	lookup := map[string]map[string]map[string]string{}
+
+	// Construct a list of all audiences except "" and "all"
+	concreteAudiences := []string{}
+	isWildcard := map[string]bool{"": true, "all": true}
+	for audience := range validAudiences {
+		if isWildcard[audience] {
+			continue
+		}
+		concreteAudiences = append(concreteAudiences, audience)
+	}
+	log.WithFields(log.Fields{
+		"concreteAudiences": concreteAudiences,
+		"isWildcard":        isWildcard,
+	}).Debug("constructing variable lookup table")
+
+	// setValue expands wildcard audiences into concrete ones.
+	var setValue func(audience, platform, name, value string)
+	setValue = func(audience, platform, name, value string) {
+		if isWildcard[audience] {
+			for _, aud := range concreteAudiences {
+				setValue(aud, platform, name, value)
+			}
+			return
+		}
+
+		if lookup[audience] == nil {
+			lookup[audience] = map[string]map[string]string{}
+		}
+		if lookup[audience][platform] == nil {
+			lookup[audience][platform] = map[string]string{}
+		}
+		log.WithFields(log.Fields{
+			"audience": audience,
+			"platform": platform,
+			"name":     name,
+			"value":    value,
+		}).Debug("setting variable")
+		lookup[audience][platform][name] = value
+	}
+
+	// Construct the lookup table for each audience+platform+name
+	for name, variable := range c.Variables {
+		log.WithFields(log.Fields{
+			"name":     name,
+			"variable": variable,
+		}).Debug("handling variable")
+		for _, value := range variable.Values {
+
+			// Two-way values should not end in path separator.
+			// Given a variable 'apps' with value '/path/to/apps',
+			// '/path/to/apps/blender' should be remapped to '{apps}/blender'.
+			if variable.Direction == "twoway" {
+				if strings.Contains(value.Value, "\\") {
+					log.WithFields(log.Fields{
+						"variable": name,
+						"value":    value,
+					}).Warning("Backslash found in variable value. Change paths to use forward slashes instead.")
+				}
+				value.Value = strings.TrimRight(value.Value, "/")
+			}
+
+			if value.Platform != "" {
+				setValue(value.Audience, value.Platform, name, value.Value)
+			}
+			for _, platform := range value.Platforms {
+				setValue(value.Audience, platform, name, value.Value)
+			}
+		}
+	}
+	log.WithFields(log.Fields{
+		"variables": c.Variables,
+		"lookup":    lookup,
+	}).Debug("constructed lookup table")
+	c.VariablesLookup = lookup
+}
+
+// ExpandVariables converts "{variable name}" to the value that belongs to the given audience and platform.
+func (c *Conf) ExpandVariables(valueToExpand, audience, platform string) string {
+	audienceMap := c.VariablesLookup[audience]
+	if audienceMap == nil {
+		log.WithFields(log.Fields{
+			"valueToExpand": valueToExpand,
+			"audience":      audience,
+			"platform":      platform,
+		}).Warning("no variables defined for this audience")
+		return valueToExpand
+	}
+
+	platformMap := audienceMap[platform]
+	if platformMap == nil {
+		log.WithFields(log.Fields{
+			"valueToExpand": valueToExpand,
+			"audience":      audience,
+			"platform":      platform,
+		}).Warning("no variables defined for this platform given this audience")
+		return valueToExpand
+	}
+
+	// Variable replacement
+	for varname, varvalue := range platformMap {
+		placeholder := fmt.Sprintf("{%s}", varname)
+		valueToExpand = strings.Replace(valueToExpand, placeholder, varvalue, -1)
+	}
+
+	return valueToExpand
+}
+
+// checkVariables performs some basic checks on variable definitions.
+// Note that the returned error only reflects the last-found error.
+// All errors are logged, though.
+func (c *Conf) checkVariables() error {
+	var err error
+
+	directionNames := []string{"oneway", "twoway"}
+	validDirections := map[string]bool{}
+	for _, direction := range directionNames {
+		validDirections[direction] = true
+	}
+
+	for name, variable := range c.Variables {
+		if !validDirections[variable.Direction] {
+			log.WithFields(log.Fields{
+				"name":      name,
+				"direction": variable.Direction,
+			}).Errorf("variable has invalid direction, choose from %v", directionNames)
+			err = ErrBadDirection
+		}
+		for valueIndex, value := range variable.Values {
+			// No platforms at all.
+			if value.Platform == "" && len(value.Platforms) == 0 {
+				log.WithFields(log.Fields{
+					"name":  name,
+					"value": value,
+				}).Error("variable has a platformless value")
+
+				err = ErrMissingVariablePlatform
+				continue
+			}
+
+			// Both Platform and Platforms.
+			if value.Platform != "" && len(value.Platforms) > 0 {
+				log.WithFields(log.Fields{
+					"name":      name,
+					"value":     value,
+					"platform":  value.Platform,
+					"platforms": value.Platforms,
+				}).Warning("variable has a both 'platform' and 'platforms' set")
+				value.Platforms = append(value.Platforms, value.Platform)
+				value.Platform = ""
+				variable.Values[valueIndex] = value
+			}
+		}
+	}
+
+	return err
 }
 
 func (c *Conf) checkDatabase() {
@@ -312,55 +558,6 @@ func (c *Conf) HasTLS() bool {
 	return c.TLSCert != "" && c.TLSKey != ""
 }
 
-// processVariables takes by-varname and turns it into by-platform, and checks for duplicates
-// between regular and path-replacement variables.
-func (c *Conf) processVariables() error {
-	foundDuplicate := false
-	for varname, perplatform := range c.PathReplacementByVarname {
-		// Check variable/path replacement duplicates.
-		_, found := c.VariablesByVarname[varname]
-		if found {
-			log.WithField("variable", varname).Error("Variable defined as both regular and path replacement variable")
-			foundDuplicate = true
-		}
-
-		// Remove trailing slashes from replacement paths, since there should be a slash after
-		// each path replacement variable anyway.
-		for platform, value := range perplatform {
-			if strings.Contains(value, "\\") {
-				log.WithFields(log.Fields{
-					"variable": varname,
-					"platform": platform,
-					"value":    value,
-				}).Warning("Backslash found in path replacement variable. Change those to forward slashes instead.")
-			}
-			perplatform[platform] = strings.TrimRight(value, "/")
-		}
-	}
-
-	transposeVariableMatrix(&c.VariablesByVarname, &c.VariablesByPlatform)
-	transposeVariableMatrix(&c.PathReplacementByVarname, &c.PathReplacementByPlatform)
-
-	for platform, vars := range c.VariablesByPlatform {
-		log.Debugf("Variables for '%s'", platform)
-		for name, value := range vars {
-			log.Debugf("     %15s = %s", name, value)
-		}
-	}
-
-	for platform, vars := range c.PathReplacementByPlatform {
-		log.Debugf("Paths for '%s'", platform)
-		for name, value := range vars {
-			log.Debugf("     %15s = %s", name, value)
-		}
-	}
-
-	if foundDuplicate {
-		return ErrDuplicateVariables
-	}
-	return nil
-}
-
 // OverrideMode checks the mode parameter for validity and logs that it's being overridden.
 func (c *Conf) OverrideMode(mode string) {
 	if mode == c.Mode {
@@ -411,18 +608,6 @@ func (c *Conf) parseURLs() {
 		}).Error("unable to construct URL to get JWT public keys")
 	} else {
 		c.JWT.PublicKeysURL = jwtURL.String()
-	}
-}
-
-func transposeVariableMatrix(in, out *map[string]map[string]string) {
-	*out = make(map[string]map[string]string)
-	for varname, perplatform := range *in {
-		for platform, varvalue := range perplatform {
-			if (*out)[platform] == nil {
-				(*out)[platform] = make(map[string]string)
-			}
-			(*out)[platform][varname] = varvalue
-		}
 	}
 }